// resources/frontend/exchanger-bot/src/app/core/features/admin/admin.info.model.ts
export interface IAdminOperationsCountInfoInDto {
  operationsCount: number;
  operationsCountToday: number;
}

export interface IAdminOperationsCountInfoOutDto {
  operations_count: number;
  operations_count_today: number;
}

export interface IAdminOperationsSumInfoInDto {
  operationsSum: number;
  operationsSumToday: number;
}

export interface IAdminOperationsSumInfoOutDto {
  operations_sum: number;
  operations_sum_today: number;
}

export interface IAdminUsersCountInfoInDto {
  usersCount: number;
  usersCountToday: number;
}

export interface IAdminUsersCountInfoOutDto {
  users_count: number;
  users_count_today: number;
}